Split business logic
