+++ www/index.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>player</title>
-    <link href='/public/player.css' rel='stylesheet' type='text/css'>
+    <link href='/player/player.css' rel='stylesheet' type='text/css'>
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
     <!--<script defer src="https://use.fontawesome.com/releases/v5.0.1/js/all.js"></script>-->
     <style>
@@ -44,7 +44,7 @@
 <video autoplay muted data-namespace="fourteen" data-controls="startstop|fullscreen|snapshot|cycle"></video>
 
 <script src ="/socket.io/socket.io.js"></script>
-<script src="/public/player.min.js">//script runs and builds mse players on top of video elements that have valid data-namespace attribute</script>
+<script src="/player/player.min.js">//script runs and builds mse players on top of video elements that have valid data-namespace attribute</script>
 
 </body>
 </html>
